refactor: 불필요한 cartPage Wrapper 삭제 및 태그 삭제
