+++ src/pages/OrderConfirmPage/OrderConfirmDataLoader.tsx
@@ -1,8 +1,7 @@
-import { cartItemSelector, cartItemsState } from "@/recoil/cartItems";
+import { cartItemsState } from "@/recoil/cartItems";
 import { useRecoilValue } from "recoil";
 import OrderConfirmPage from "./OrderConfirmPage";
 import { useEffect, useState } from "react";
-import CartPageSkeleton from "../CartPage/CartPage.skeleton";
 import MainLayout from "@/components/layout/MainLayout";
 import BackButton from "@/components/_common/BackButton/BackButton";
 import { useNavigate } from "react-router-dom";
@@ -10,15 +9,14 @@ import { CART_PAGE_MESSAGES } from "@/constants/cart";
 import { CartItem } from "@/types/cart";
 import { ErrorBoundary } from "react-error-boundary";
 import ErrorPage from "../Error/ErrorPage";
-import { getLocalStorage } from "@/utils/localStorage";
 import { PAGE_URL } from "@/constants/url";
 import { ErrorMessage } from "@/constants/error";
+import { selectedCartItemList } from "@/recoil/selectedCardItems";
 
 const OrderConfirmDataLoader = () => {
-  const cartItems = useRecoilValue(cartItemSelector);
   const cartItem = useRecoilValue(cartItemsState);
   const [selectedItems, setSelectedItems] = useState<CartItem[]>([]);
-
+  const selectedCartItems = useRecoilValue(selectedCartItemList);
   const navigate = useNavigate();
 
   const onMoveCartPage = () => {
@@ -28,18 +26,13 @@ const OrderConfirmDataLoader = () => {
   };
 
   useEffect(() => {
-    const selectedItemsId = getLocalStorage("selectedItems");
-    const selectedCartItems = cartItems.filter((item) =>
-      selectedItemsId.includes(item.id)
-    );
-
     setSelectedItems(selectedCartItems);
 
-    if (!selectedItemsId.length) {
+    if (!selectedCartItems.length) {
       alert(ErrorMessage.noOrderInfo);
       navigate(PAGE_URL.home);
     }
-  }, [cartItems, navigate]);
+  }, [selectedCartItems, navigate]);
 
   return (
     <MainLayout>
@@ -47,7 +40,7 @@ const OrderConfirmDataLoader = () => {
         <BackButton onClick={onMoveCartPage} />
       </MainLayout.Header>
       <ErrorBoundary FallbackComponent={ErrorPage}>
-        <MainLayout.Body fallback={<CartPageSkeleton />}>
+        <MainLayout.Body>
           {cartItem.length > 0 && (
             <OrderConfirmPage selectedCartItems={selectedItems} />
           )}
