+++ src/recoil/selectedCardItems.ts
@@ -1,5 +1,7 @@
-import { setLocalStorage } from "@/utils/localStorage";
-import { AtomEffect, atom } from "recoil";
+import { cartItemsState } from "@/recoil/cartItems";
+import { CartItem } from "@/types/cart";
+import { getLocalStorage, setLocalStorage } from "@/utils/localStorage";
+import { AtomEffect, atom, selector } from "recoil";
 
 const localStorageEffectForItem =
   (): AtomEffect<number[]> =>
@@ -13,4 +15,15 @@ export const selectedCartItemsIdState = atom<number[]>({
   key: "selectedCartItems",
   default: [],
   effects: [localStorageEffectForItem()],
+});
+
+export const selectedCartItemList = selector<CartItem[]>({
+  key: "selectedCartItemList",
+  get: ({ get }) => {
+    const selectedItemsId = getLocalStorage("selectedItems");
+    const selectedCartItems = get(cartItemsState).filter((item) =>
+      selectedItemsId.includes(item.id)
+    );
+    return selectedCartItems;
+  },
 });
